Added main methods to InternMethodDemo and Palindrome classes
In this update, we've added a main method to both the InternMethodDemo and Palindrome classes. The InternMethodDemo's main method demonstrates the use of Java's intern() method for string comparison. In the Palindrome class, we've implemented a new isPalindrome function that checks if a given string is a palindrome or not, and used it in the main method.