Added .gitignore and pom.xml files
A new .gitignore file has been added to ignore unnecessary files and folders from various IDEs, build tools, and operating systems. This includes IntelliJ IDEA, Eclipse, NetBeans, VS Code, Maven target directories, and Mac OS specific files.

Additionally, a new pom.xml file has been created for the Maven project configuration. It specifies the group ID, artifact ID and version of the project along with properties such as compiler source/target versions and project build source encoding.

Added basic string and object operations

In this update, we've introduced some fundamental operations with strings and objects. This includes creating a greeting message, calculating the length of a string, using String.format for personalized greetings, and working with Object's hashCode method. We also demonstrated the use of equals method to compare objects.

Added new StringPoolDemo class

A new class named StringPoolDemo has been introduced. Currently, it's an empty class without any methods or properties.

Added main method to StringPoolDemo

The main method was added to the StringPoolDemo class. This includes the creation of three string variables and two print statements that compare these strings using the equals() method.

Refactor string comparison to hash code printing

The previous implementation was comparing strings using the equals method. This has been refactored to print out the hash codes of the strings instead. A new helper function, printHashCode, has been introduced for this purpose. It takes a variable name and a string as parameters, calculates the hash code of the string, and prints it along with the variable name.

Added two new classes

Two new empty classes have been created. These will be used for future development and feature implementation.

Added main methods to InternMethodDemo and Palindrome classes

In this update, we've added a main method to both the InternMethodDemo and Palindrome classes. The InternMethodDemo's main method demonstrates the use of Java's intern() method for string comparison. In the Palindrome class, we've implemented a new isPalindrome function that checks if a given string is a palindrome or not, and used it in the main method.