Added age calculation and hands count
In the updated code, a new method for calculating age has been introduced in the Person class. This method takes a birth year as input and prints out the calculated age. Additionally, a static variable representing the number of hands has been added to the Person class. These changes have also been reflected in Main.java where IronMan's age is now calculated using this new method and the number of hands is printed out.